ubik: switch to 64-bit version
Switch the ubik code internally to use the new version structure
with 64-bit members.  Keep the old version format for now for the
interface and convert when necessary.

+++ src/kauth/rebuild.c
@@ -63,8 +63,8 @@ readUbikHeader(void)
 	printf("Ubik Header\n");
 	printf("   Magic           = 0x%x\n", uheader.magic);
 	printf("   Size            = %u\n", uheader.size);
-	printf("   Version.epoch   = %u\n", uheader.version.epoch);
-	printf("   Version.counter = %u\n", uheader.version.counter);
+	printf("   Version.epoch   = %lu\n", (long)uheader.version.epoch);
+	printf("   Version.counter = %lu\n", (long)uheader.version.counter);
     }
 
     if (uheader.size != UBIK_HEADERSIZE)
